Migrate rankables#sort out of admin namespace
